work towards conditionally rendering Signup vs Login
