+++ react_frontend/src/welcome/welcome.js
@@ -1,53 +1,56 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { createUser } from '../store/user/actions';
+// import { createUser } from '../store/user/actions';
+import Signup from './Signup';
+import Login from './Login';
 
 class Welcome extends Component{
     state = {
+        signup: false,
         loggedIn: this.props.user.loggedIn,
-        user: {...this.props.user}
+        user: {...this.props.user.userInfo}
     }
 
-    handleUsernameInput = (e) => {
-        // console.log("username: " + e.target.value)
-        this.setState({
-            user: {
-                user_name: e.target.value,
-                password: this.state.user.password,
-                email: this.state.user.email
-            }
-        })
-    }
+    // handleUsernameInput = (e) => {
+    //     // console.log("username: " + e.target.value)
+    //     this.setState({
+    //         user: {
+    //             user_name: e.target.value,
+    //             password: this.state.user.password,
+    //             email: this.state.user.email
+    //         }
+    //     })
+    // }
 
-    handleEmailInput = (e) => {
-        // console.log("Email: " + e.target.value)
-        this.setState({
-            user: {
-                user_name: this.state.user.user_name,
-                password: this.state.user.password,
-                email: e.target.value
-            }
-        })
-    }
+    // handleEmailInput = (e) => {
+    //     // console.log("Email: " + e.target.value)
+    //     this.setState({
+    //         user: {
+    //             user_name: this.state.user.user_name,
+    //             password: this.state.user.password,
+    //             email: e.target.value
+    //         }
+    //     })
+    // }
 
-    handlePasswordInput = (e) => {
-        // console.log("Password: " + e.target.value)
-        this.setState({
-            user: {
-                user_name: this.state.user.user_name,
-                password: e.target.value,
-                email: this.state.user.email
-            }
-        })
-    }
+    // handlePasswordInput = (e) => {
+    //     // console.log("Password: " + e.target.value)
+    //     this.setState({
+    //         user: {
+    //             user_name: this.state.user.user_name,
+    //             password: e.target.value,
+    //             email: this.state.user.email
+    //         }
+    //     })
+    // }
 
-    handleSubmit = (e) => {
-        e.preventDefault()
-        console.log(e.target)
+    // handleSubmit = (e) => {
+    //     e.preventDefault()
+    //     console.log(e.target)
 
-        this.props.handleCreateUser({...this.state.user})
+    //     this.props.handleCreateUser({...this.state.user})
 
-    }
+    // }
 
 
     render() {
@@ -61,8 +64,9 @@ class Welcome extends Component{
             <div>
                 <h1>Welcome to Three Seeds Tarot App</h1>
 
+                { this.state.signup ? <Signup/> : <Login/> }
 
-                <form onSubmit={this.handleSubmit}> 
+                {/* <form onSubmit={this.handleSubmit}> 
                     <label>
                         Username:
                     </label>
@@ -79,7 +83,7 @@ class Welcome extends Component{
                     <input type ="password" name="password" placeholder="password" onChange={this.handlePasswordInput}/>
 
                     <button type='submit' name="Submit"/>
-                </form>
+                </form> */}
             </div>
     )}
 }
@@ -88,10 +92,10 @@ const mapStateToProps = (state) => ({
     user: {...state.user}
 })
 
-const mapDispatchToProps = (dispatch) => ({
-    handleCreateUser: user => {
-        dispatch(createUser(user))
-    }
-})
+// const mapDispatchToProps = (dispatch) => ({
+//     handleCreateUser: user => {
+//         dispatch(createUser(user))
+//     }
+// })
 
-export default connect(mapStateToProps, mapDispatchToProps)(Welcome)
+export default connect(mapStateToProps)(Welcome)
